llamA/project/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
docx2txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
pydantic
requests
